+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 BREAKING_CHANGES:  
 - Ability to override only one template (issue #166, thanks @Nihisil)  
+- removed `TPromise` type for `--responses` options (perf. problem, issue #182, thanks @mixalbl4-127)  
 
 Features:  
 - template path prefixes `@base`, `@default`, `@modular` (using in Eta templates, `includeFile()`, see README.md)