+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # next release  
 
+# 6.0.0  
+
 BREAKING_CHANGES:  
 - Ability to override only one template (issue #166, thanks @Nihisil)  
 - removed `TPromise` type for `--responses` options (perf. problem, issue #182, thanks @mixalbl4-127)